requirements.txt
Flask==3.0.2
pypdf>=5.1.0
openai>=1.81.0
PyYAML>=6.0.0